# docs/architecture/openapi/admin-v1.yaml
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Happy Agent Platform Admin API",
    "version": "1.0.0",
    "description": "Closed control-plane contract for typed component catalogs, defaults, Agent lifecycle, evaluation, publication, sessions, runs, traces and playground."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin application"
    }
  ],
  "security": [
    {
      "agentAdminBearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/admin/frameworks": {
      "get": {
        "tags": [
          "Frameworks"
        ],
        "operationId": "listFrameworks",
        "summary": "List registered frameworks",
        "description": "Lists code-registered framework adapter versions and capabilities.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/ComponentStatusFilter"
          },
          {
            "$ref": "#/components/parameters/ComponentCategoryFilter"
          },
          {
            "$ref": "#/components/parameters/CompatibleFrameworkFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Framework page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameworkPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/admin/providers": {
      "get": {
        "tags": [
          "Providers"
        ],
        "operationId": "listProviders",
        "summary": "List Provider versions",
        "description": "Lists typed Provider versions with masked credentials and health.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/ComponentStatusFilter"
          },
          {
            "$ref": "#/components/parameters/ComponentCategoryFilter"
          },
          {
            "$ref": "#/components/parameters/CompatibleFrameworkFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "tags": [
          "Providers"
        ],
        "operationId": "createProvider",
        "summary": "Create a Provider version",
        "description": "Creates a typed Provider version and encrypts the supplied credential.",
        "x-error-codes": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "IDEMPOTENCY_CONFLICT",
          "CONFLICT"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderWriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created Provider",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderComponent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/admin/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "operationId": "listModels",
        "summary": "List Models",
        "description": "Lists typed model configurations and discovered capabilities.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/ComponentStatusFilter"
          },
          {
            "$ref": "#/components/parameters/ComponentCategoryFilter"
          },
          {
            "$ref": "#/components/parameters/CompatibleFrameworkFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Model page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "tags": [
          "Models"
        ],
        "operationId": "createModel",
        "summary": "Create a Model",
        "description": "Creates a typed Model bound to one Provider version.",
        "x-error-codes": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "IDEMPOTENCY_CONFLICT",
          "CONFLICT"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelWriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created Model",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelComponent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/admin/models/{modelKey}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ModelKey"
        }
      ],
      "get": {
        "tags": [
          "Models"
        ],
        "operationId": "getModel",
        "summary": "Read a Model",
        "description": "Returns Model capabilities, parameter schema, version diff and usage impact.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND"
        ],
        "responses": {
          "200": {
            "description": "Model",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelComponent"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "patch": {
        "tags": [
          "Models"
        ],
        "operationId": "updateModel",
        "summary": "Update a Model",
        "description": "Updates explicit editable Model fields with optimistic concurrency.",
        "x-optimistic-concurrency": true,
        "x-error-codes": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "PRECONDITION_FAILED",
          "PRECONDITION_REQUIRED",
          "IDEMPOTENCY_CONFLICT"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Model",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelComponent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/admin/tools": {
      "get": {
        "tags": [
          "Tools"
        ],
        "operationId": "listTools",
        "summary": "List code-registered Tools",
        "description": "Lists immutable Tool contract versions discovered from application Beans.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/ComponentStatusFilter"
          },
          {
            "$ref": "#/components/parameters/ComponentCategoryFilter"
          },
          {
            "$ref": "#/components/parameters/CompatibleFrameworkFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Tool page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/admin/skills": {
      "get": {
        "tags": [
          "Skills"
        ],
        "operationId": "listSkills",
        "summary": "List Skill versions",
        "description": "Lists versioned Markdown and metadata Skills with compatibility and impact.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/ComponentStatusFilter"
          },
          {
            "$ref": "#/components/parameters/ComponentCategoryFilter"
          },
          {
            "$ref": "#/components/parameters/CompatibleFrameworkFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Skill page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "tags": [
          "Skills"
        ],
        "operationId": "createSkill",
        "summary": "Create a Skill draft version",
        "description": "Creates a Skill containing Markdown, structured metadata, read-only resources and authorized Tool requirements.",
        "x-error-codes": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "IDEMPOTENCY_CONFLICT",
          "CONFLICT"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillWriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created Skill",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillComponent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/admin/hooks": {
      "get": {
        "tags": [
          "Hooks"
        ],
        "operationId": "listHooks",
        "summary": "List code-registered Hooks",
        "description": "Lists lifecycle Hooks, ordering, failure policy, defaults and compatibility.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/ComponentStatusFilter"
          },
          {
            "$ref": "#/components/parameters/ComponentCategoryFilter"
          },
          {
            "$ref": "#/components/parameters/CompatibleFrameworkFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Hook page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/admin/prompts": {
      "get": {
        "tags": [
          "Prompts"
        ],
        "operationId": "listPrompts",
        "summary": "List Prompt versions",
        "description": "Lists versioned prompt templates and variable schemas.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/ComponentStatusFilter"
          },
          {
            "$ref": "#/components/parameters/ComponentCategoryFilter"
          },
          {
            "$ref": "#/components/parameters/CompatibleFrameworkFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Prompt page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "tags": [
          "Prompts"
        ],
        "operationId": "createPrompt",
        "summary": "Create a Prompt version",
        "description": "Creates a Prompt draft with a closed variable schema.",
        "x-error-codes": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "IDEMPOTENCY_CONFLICT",
          "CONFLICT"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptWriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created Prompt",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptComponent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/admin/prompts": {
      "post": {
        "tags": ["Prompts"],
        "operationId": "adminCreatePrompt",
        "summary": "Create a prompt for the personal workbench",
        "description": "Creates one independently managed prompt that can be selected by an Agent draft.",
        "x-error-codes": ["VALIDATION_FAILED", "UNAUTHORIZED", "FORBIDDEN", "CONFLICT", "IDEMPOTENCY_CONFLICT"],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchPromptCreate"}}}
        },
        "responses": {
          "201": {"description": "Prompt created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchPrompt"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/Unprocessable"},
          "default": {"$ref": "#/components/responses/Problem"}
        }
      }
    },
    "/api/admin/skills": {
      "post": {
        "tags": ["Skills"],
        "operationId": "adminCreateSkill",
        "summary": "Create a skill for the personal workbench",
        "description": "Creates one independently managed skill after validating its runtime Tool dependencies.",
        "x-error-codes": ["VALIDATION_FAILED", "UNAUTHORIZED", "FORBIDDEN", "CONFLICT", "IDEMPOTENCY_CONFLICT"],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchSkillCreate"}}}
        },
        "responses": {
          "201": {"description": "Skill created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchSkill"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/Unprocessable"},
          "default": {"$ref": "#/components/responses/Problem"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "agentAdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Platform administrator token only.",
        "x-audience": "happy-agent-admin-v1",
        "x-required-scope": "AGENT_ADMIN"
      }
    },
...
    "parameters": {
...
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "ComponentStatusFilter": {
        "name": "status",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/LifecycleStatus"
        }
      },
      "ComponentCategoryFilter": {
        "name": "category",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 120
        }
      },
      "CompatibleFrameworkFilter": {
        "name": "compatibleFramework",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/FrameworkKeyValue"
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Malformed request.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    },
...
  }
}